Use float for demo position percentage.

+++ src/client/client.h
@@ -463,7 +463,7 @@ typedef struct client_static_s {
         int         last_snapshot;      // number of demo frame the last snapshot was saved
         int64_t     file_size;
         int64_t     file_offset;
-        int         file_percent;
+        float       file_percent;
         sizebuf_t   buffer;
         list_t      snapshots;
         bool        paused;
